# Optimize test_sse2 to run faster on the CI.