… test/sse/test_sse2.cpp