Before:

```
test_sse2 (test_core.core0.test_sse2) ... ok (5.79s)
test_sse2 (test_core.core1.test_sse2) ... ok (5.79s)
test_sse2 (test_core.core2.test_sse2) ... ok (9.20s)
test_sse2 (test_core.core_2gb.test_sse2) ... ok (6.14s)
test_sse2 (test_core.cores.test_sse2) ... ok (9.96s)
test_sse2 (test_core.corez.test_sse2) ... ok (8.74s)
test_sse2 (test_core.core3.test_sse2) ... ok (99.19s)
```

After:

```
test_sse2 (test_core.core0.test_sse2) ... ok (5.78s)
test_sse2 (test_core.core1.test_sse2) ... ok (6.67s)
test_sse2 (test_core.core2.test_sse2) ... ok (9.52s)
test_sse2 (test_core.core_2gb.test_sse2) ... ok (6.17s)
test_sse2 (test_core.cores.test_sse2) ... ok (9.20s)
test_sse2 (test_core.corez.test_sse2) ... ok (8.42s)
test_sse2 (test_core.core3.test_sse2) ... ok (26.85s)
```